+++ src/Deed/Deed.Application/Exchanges/Responses/ExchangeResponse.cs
@@ -3,6 +3,7 @@ namespace Deed.Application.Exchanges.Responses;
 public sealed record ExchangeResponse(
     string TargetCurrency,
     string NationalCurrency,
-    float Buy,
-    float Sale,
-    DateTimeOffset? UpdatedAt);
+    decimal Buy,
+    decimal Sale,
+    DateTimeOffset? UpdatedAt
+);
